WIP: Removed format specific Timesteps
