+++ package/MDAnalysis/coordinates/DMS.py
@@ -1,5 +1,5 @@
 # -*- Mode: python; tab-width: 4; indent-tabs-mode:nil; coding:utf-8 -*-
-# vim: tabstop=4 expandtab shiftwidth=4 softtabstop=4 
+# vim: tabstop=4 expandtab shiftwidth=4 softtabstop=4
 #
 # MDAnalysis --- https://www.mdanalysis.org
 # Copyright (c) 2006-2017 The MDAnalysis Development Team and contributors
@@ -37,28 +37,6 @@ import numpy as np
 import sqlite3
 
 from . import base
-from .core import triclinic_box, triclinic_vectors
-
-
-class Timestep(base.Timestep):
-    def _init_unitcell(self):
-        return {'x': np.zeros(3),
-                'y': np.zeros(3),
-                'z': np.zeros(3)}
-
-    @property
-    def dimensions(self):
-        """unitcell dimensions (A, B, C, alpha, beta, gamma)"""
-        x = self._unitcell['x']
-        y = self._unitcell['y']
-        z = self._unitcell['z']
-        return triclinic_box(x, y, z)
-
-    @dimensions.setter
-    def dimensions(self, box):
-        x, y, z = triclinic_vectors(box)
-        cell = {'x': x, 'y': y, 'z': z}
-        self._unitcell = cell
 
 
 class DMSReader(base.SingleFrameReaderBase):
@@ -70,7 +48,6 @@ class DMSReader(base.SingleFrameReaderBase):
     """
     format = 'DMS'
     units = {'time': None, 'length': 'A', 'velocity': 'A/ps'}
-    _Timestep = Timestep
 
     def get_coordinates(self, cur):
         cur.execute('SELECT * FROM particle')
@@ -87,11 +64,12 @@ class DMSReader(base.SingleFrameReaderBase):
     def get_global_cell(self, cur):
         cur.execute('SELECT * FROM global_cell')
         rows = cur.fetchall()
-        assert len(rows) == 3
+
         x = [row["x"] for row in rows]
         y = [row["y"] for row in rows]
         z = [row["z"] for row in rows]
-        return {'x': x, 'y': y, 'z': z}
+
+        return np.array([x, y, z], dtype=np.float32)
 
     def _read_first_frame(self):
         coords_list = None
@@ -119,16 +97,16 @@ class DMSReader(base.SingleFrameReaderBase):
         if not velocities.any():
             velocities = None
 
-        self.ts = self._Timestep.from_coordinates(
+        self.ts = base.Timestep.from_coordinates(
             np.array(coords_list, dtype=np.float32),
             velocities=velocities,
             **self._ts_kwargs)
         self.ts.frame = 0  # 0-based frame number
 
-        self.ts._unitcell = unitcell
+        self.ts.triclinic_dimensions = unitcell
         if self.convert_units:
             self.convert_pos_from_native(self.ts._pos)  # in-place !
-            self.convert_pos_from_native(self.ts._unitcell)  # in-place ! (all are lengths)
+            self.convert_pos_from_native(self.ts.dimensions[:3])  # in-place ! (all are lengths)
             if self.ts.has_velocities:
                 # converts nm/ps to A/ps units
                 self.convert_velocities_from_native(self.ts._velocities)
