+++ package/MDAnalysis/coordinates/LAMMPS.py
@@ -192,7 +192,7 @@ class DATAReader(base.SingleFrameReaderBase):
 
     def _read_first_frame(self):
         with DATAParser(self.filename) as p:
-            self.ts = p.read_DATA_timestep(self.n_atoms, self._Timestep,
+            self.ts = p.read_DATA_timestep(self.n_atoms, base.Timestep,
                                            self._ts_kwargs)
 
         self.ts.frame = 0
