+++ package/MDAnalysis/coordinates/TXYZ.py
@@ -63,7 +63,6 @@ class TXYZReader(base.ReaderBase):
     format = ['TXYZ', 'ARC']
     # these are assumed!
     units = {'time': 'ps', 'length': 'Angstrom'}
-    _Timestep = base.Timestep
 
     def __init__(self, filename, **kwargs):
         super(TXYZReader, self).__init__(filename, **kwargs)
@@ -75,7 +74,7 @@ class TXYZReader(base.ReaderBase):
         self.xyzfile = util.anyopen(self.filename)
         self._cache = dict()
 
-        self.ts = self._Timestep(self.n_atoms, **self._ts_kwargs)
+        self.ts = base.Timestep(self.n_atoms, **self._ts_kwargs)
         # Haven't quite figured out where to start with all the self._reopen()
         # etc.
         # (Also cannot just use seek() or reset() because that would break
